streamlit_template/requirements.txt
black==24.3.0
isort==5.13.2
numpy==1.26.3
streamlit==1.36.0
pandas==2.2.2
plotly==5.22.0
pytest==8.3.2
pre-commit==3.6.0
snowflake-connector-python==3.9.1
snowflake-snowpark-python==1.15.0
